refactor(service): split startup orchestration helpers
Move service startup, clock monitor, media history, publisher, and indexing helpers into focused files. Stabilize notification wait tests by synchronizing broadcasts on WebSocket session readiness instead of fixed sleeps.
