+++ pkg/service/service.go
@@ -22,24 +22,14 @@ along with Zaparoo Core.  If not, see <http://www.gnu.org/licenses/>.
 package service
 
 import (
-	"context"
 	"fmt"
-	"os"
 	"path/filepath"
 	"sync"
-	"time"
 
 	"github.com/ZaparooProject/zaparoo-core/v2/pkg/api"
-	"github.com/ZaparooProject/zaparoo-core/v2/pkg/api/methods"
 	"github.com/ZaparooProject/zaparoo-core/v2/pkg/api/models"
-	"github.com/ZaparooProject/zaparoo-core/v2/pkg/api/notifications"
 	"github.com/ZaparooProject/zaparoo-core/v2/pkg/audio"
 	"github.com/ZaparooProject/zaparoo-core/v2/pkg/config"
-	"github.com/ZaparooProject/zaparoo-core/v2/pkg/database"
-	"github.com/ZaparooProject/zaparoo-core/v2/pkg/database/mediadb"
-	"github.com/ZaparooProject/zaparoo-core/v2/pkg/database/systemdefs"
-	"github.com/ZaparooProject/zaparoo-core/v2/pkg/database/userdb"
-	"github.com/ZaparooProject/zaparoo-core/v2/pkg/database/userdb/boltmigration"
 	"github.com/ZaparooProject/zaparoo-core/v2/pkg/groovyproxy"
 	"github.com/ZaparooProject/zaparoo-core/v2/pkg/helpers"
 	"github.com/ZaparooProject/zaparoo-core/v2/pkg/helpers/syncutil"
@@ -50,230 +40,14 @@ import (
 	"github.com/ZaparooProject/zaparoo-core/v2/pkg/service/inbox"
 	"github.com/ZaparooProject/zaparoo-core/v2/pkg/service/playlists"
 	"github.com/ZaparooProject/zaparoo-core/v2/pkg/service/playtime"
-	"github.com/ZaparooProject/zaparoo-core/v2/pkg/service/publishers"
 	"github.com/ZaparooProject/zaparoo-core/v2/pkg/service/state"
 	"github.com/ZaparooProject/zaparoo-core/v2/pkg/service/tokens"
 	"github.com/ZaparooProject/zaparoo-core/v2/pkg/service/updater"
 	"github.com/ZaparooProject/zaparoo-core/v2/pkg/zapscript"
 	"github.com/google/uuid"
 	"github.com/jonboulle/clockwork"
-	"github.com/mackerelio/go-osstat/uptime"
 	"github.com/rs/zerolog/log"
 )
-
-const zapLinkHostExpiration = 30 * 24 * time.Hour
-
-func setupEnvironment(pl platforms.Platform) error {
-	if _, ok := helpers.HasUserDir(); ok {
-		log.Info().Msg("using 'user' directory for storage")
-	}
-
-	log.Info().Msg("creating platform directories")
-	dirs := []string{
-		helpers.ConfigDir(pl),
-		pl.Settings().TempDir,
-		helpers.DataDir(pl),
-		filepath.Join(helpers.DataDir(pl), config.MappingsDir),
-		filepath.Join(helpers.DataDir(pl), config.AssetsDir),
-		filepath.Join(helpers.DataDir(pl), config.LaunchersDir),
-		filepath.Join(helpers.DataDir(pl), config.MediaDir),
-	}
-	for _, dir := range dirs {
-		err := os.MkdirAll(dir, 0o750)
-		if err != nil {
-			return fmt.Errorf("failed to create directory %s: %w", dir, err)
-		}
-	}
-
-	return nil
-}
-
-func makeDatabase(ctx context.Context, pl platforms.Platform) (*database.Database, error) {
-	db := &database.Database{
-		MediaDB: nil,
-		UserDB:  nil,
-	}
-
-	log.Debug().Msg("opening media database")
-	mediaDB, err := mediadb.OpenMediaDB(ctx, pl)
-	if err != nil {
-		return db, fmt.Errorf("failed to open media database: %w", err)
-	}
-
-	log.Debug().Msg("running media database migrations")
-	err = mediaDB.MigrateUp()
-	if err != nil {
-		return db, fmt.Errorf("error migrating mediadb: %w", err)
-	}
-
-	db.MediaDB = mediaDB
-
-	log.Debug().Msg("opening user database")
-	userDB, err := userdb.OpenUserDB(ctx, pl)
-	if err != nil {
-		return db, fmt.Errorf("failed to open user database: %w", err)
-	}
-
-	log.Debug().Msg("running user database migrations")
-	err = userDB.MigrateUp()
-	if err != nil {
-		return db, fmt.Errorf("error migrating userdb: %w", err)
-	}
-
-	db.UserDB = userDB
-
-	// migrate old boltdb mappings if required
-	log.Debug().Msg("checking for boltdb migration")
-	err = boltmigration.MaybeMigrate(pl, userDB)
-	if err != nil {
-		log.Error().Err(err).Msg("error migrating old boltdb mappings")
-	}
-
-	return db, nil
-}
-
-func closeDatabase(db *database.Database) {
-	if db == nil {
-		return
-	}
-	if db.UserDB != nil {
-		if err := db.UserDB.Close(); err != nil {
-			log.Warn().Err(err).Msg("error closing user database")
-		}
-	}
-	if db.MediaDB != nil {
-		if err := db.MediaDB.Close(); err != nil {
-			log.Warn().Err(err).Msg("error closing media database")
-		}
-	}
-}
-
-func startupMaintenanceCancelled(ctx context.Context, message string) bool {
-	if err := ctx.Err(); err != nil {
-		log.Debug().Err(err).Msg(message)
-		return true
-	}
-	return false
-}
-
-func cleanupHistoryRetention(ctx context.Context, cfg *config.Instance, db *database.Database) {
-	if startupMaintenanceCancelled(ctx, "skipping history retention cleanup: startup maintenance cancelled") {
-		return
-	}
-	if db == nil {
-		log.Warn().Msg("skipping history retention cleanup: database is nil")
-		return
-	}
-	if db.UserDB == nil {
-		log.Warn().Msg("skipping history retention cleanup: user database is nil")
-		return
-	}
-
-	scanHistoryDays := cfg.ScanHistory()
-	if scanHistoryDays > 0 {
-		log.Info().Msgf("cleaning up scan history older than %d days", scanHistoryDays)
-		rowsDeleted, cleanupErr := db.UserDB.CleanupHistory(scanHistoryDays)
-		switch {
-		case cleanupErr != nil:
-			log.Error().Err(cleanupErr).Msg("error cleaning up scan history")
-		case rowsDeleted > 0:
-			log.Info().Msgf("deleted %d old scan history entries", rowsDeleted)
-		default:
-			log.Debug().Msg("no old scan history entries to clean up")
-		}
-	} else {
-		log.Debug().Msg("scan history cleanup disabled (retention set to 0)")
-	}
-
-	if startupMaintenanceCancelled(ctx, "skipping media history retention cleanup: startup maintenance cancelled") {
-		return
-	}
-
-	// Cleanup old media history entries if retention is configured
-	playtimeRetention := cfg.PlaytimeRetention()
-	if playtimeRetention > 0 {
-		log.Info().Msgf("cleaning up media history older than %d days", playtimeRetention)
-		rowsDeleted, cleanupErr := db.UserDB.CleanupMediaHistory(playtimeRetention)
-		switch {
-		case cleanupErr != nil:
-			log.Error().Err(cleanupErr).Msg("error cleaning up media history")
-		case rowsDeleted > 0:
-			log.Info().Msgf("deleted %d old media history entries", rowsDeleted)
-		default:
-			log.Debug().Msg("no old media history entries to clean up")
-		}
-	} else {
-		log.Debug().Msg("media history cleanup disabled (retention set to 0)")
-	}
-}
-
-func closeHangingMediaHistoryOnStartup(db *database.Database) {
-	log.Info().Msg("closing hanging media history entries")
-	if hangingErr := db.UserDB.CloseHangingMediaHistory(); hangingErr != nil {
-		log.Error().Err(hangingErr).Msg("error closing hanging media history entries")
-	}
-}
-
-// pruneExpiredZapLinkHosts removes non-supporting zaplink hosts older than 30 days.
-// This allows hosts that may have added zaplink support to be re-checked.
-func pruneExpiredZapLinkHosts(db *database.Database) {
-	log.Info().Msg("pruning expired non-supporting zaplink hosts")
-	rowsDeleted, err := db.UserDB.PruneExpiredZapLinkHosts(zapLinkHostExpiration)
-	switch {
-	case err != nil:
-		log.Error().Err(err).Msg("error pruning expired zaplink hosts")
-	case rowsDeleted > 0:
-		log.Info().Msgf("pruned %d expired non-supporting zaplink hosts", rowsDeleted)
-	default:
-		log.Debug().Msg("no expired zaplink hosts to prune")
-	}
-}
-
-func runMediaDBStartupMaintenance(ctx context.Context, db database.MediaDBI) {
-	if db == nil {
-		log.Warn().Msg("skipping media database startup maintenance: media database is nil")
-		return
-	}
-
-	db.TrackBackgroundOperation()
-	defer db.BackgroundOperationDone()
-
-	if sqlDB := db.UnsafeGetSQLDb(); sqlDB != nil {
-		log.Debug().Msg("running media database PRAGMA optimize")
-		if _, err := sqlDB.ExecContext(ctx, "PRAGMA optimize;"); err != nil {
-			log.Warn().Err(err).Msg("failed to run PRAGMA optimize")
-		}
-
-		log.Debug().Msg("running media database WAL checkpoint")
-		if _, err := sqlDB.ExecContext(ctx, "PRAGMA wal_checkpoint(TRUNCATE);"); err != nil {
-			log.Warn().Err(err).Msg("failed to run WAL checkpoint on startup")
-		}
-	} else {
-		log.Warn().Msg("skipping media database PRAGMA maintenance: SQL database is nil")
-	}
-
-	if startupMaintenanceCancelled(ctx, "skipping tag cache warmup: startup maintenance cancelled") {
-		return
-	}
-
-	if err := db.RebuildTagCache(); err != nil {
-		log.Warn().Err(err).Msg("failed to warm tag cache on startup")
-	}
-}
-
-func runStartupMaintenance(ctx context.Context, cfg *config.Instance, db *database.Database) {
-	if db == nil {
-		log.Warn().Msg("skipping startup maintenance: database is nil")
-		return
-	}
-
-	runMediaDBStartupMaintenance(ctx, db.MediaDB)
-	cleanupHistoryRetention(ctx, cfg, db)
-	if startupMaintenanceCancelled(ctx, "skipping zaplink host pruning: startup maintenance cancelled") {
-		return
-	}
-	pruneExpiredZapLinkHosts(db)
-}
 
 // StartResult holds the return values from Start.
 type StartResult struct {
@@ -570,371 +344,4 @@ func Start(
 		Done:             doneCh,
 		RestartRequested: st.RestartRequested,
 	}, nil
-}
-
-// monitorClockAndHealTimestamps monitors the system clock and heals timestamps when NTP syncs.
-// This is critical for MiSTer devices that boot without RTC and initially show 1970 epoch time.
-// Once NTP syncs, we can mathematically reconstruct correct timestamps using monotonic uptime.
-func monitorClockAndHealTimestamps(ctx context.Context, db *database.Database, bootUUID string) {
-	ticker := time.NewTicker(1 * time.Minute)
-	defer ticker.Stop()
-
-	healed := false
-	wasReliable := helpers.IsClockReliable(time.Now())
-
-	for {
-		select {
-		case <-ticker.C:
-			now := time.Now()
-			isReliable := helpers.IsClockReliable(now)
-
-			// Detect transition from unreliable → reliable (NTP sync event)
-			if !wasReliable && isReliable && !healed {
-				log.Info().Msg("clock became reliable (NTP sync detected), healing timestamps")
-
-				// Calculate true boot time: Current Time - System Uptime
-				systemUptime, err := uptime.Get()
-				if err != nil {
-					log.Error().Err(err).Msg("failed to get system uptime for timestamp healing")
-					wasReliable = isReliable
-					continue
-				}
-
-				trueBootTime := now.Add(-systemUptime)
-				log.Info().
-					Time("true_boot_time", trueBootTime).
-					Dur("uptime", systemUptime).
-					Msg("calculated true boot time")
-
-				// Heal all timestamps for this boot session
-				rowsHealed, healErr := db.UserDB.HealTimestamps(bootUUID, trueBootTime)
-				if healErr != nil {
-					log.Error().Err(healErr).Msg("failed to heal timestamps")
-				} else if rowsHealed > 0 {
-					log.Info().Int64("rows", rowsHealed).Msg("successfully healed timestamps")
-				}
-
-				healed = true
-			}
-
-			wasReliable = isReliable
-
-		case <-ctx.Done():
-			return
-		}
-	}
-}
-
-// mediaHistoryTracker encapsulates the state and logic for tracking media history.
-// It coordinates between the notification listener and the periodic PlayTime updater.
-type mediaHistoryTracker struct {
-	clock                     clockwork.Clock
-	currentMediaStartTime     time.Time
-	currentMediaStartTimeMono time.Time
-	st                        *state.State
-	db                        *database.Database
-	currentHistoryDBID        int64
-	mu                        syncutil.RWMutex
-}
-
-// listen processes media start/stop notifications and records them in the database.
-func (t *mediaHistoryTracker) listen(notificationChan <-chan models.Notification) {
-	for notif := range notificationChan {
-		switch notif.Method {
-		case models.NotificationStarted:
-			// Media started - create new history entry
-			activeMedia := t.st.ActiveMedia()
-			if activeMedia != nil {
-				now := t.clock.Now()
-				nowMono := time.Now() // Monotonic clock for duration calculation
-
-				// Calculate system uptime for timestamp healing on MiSTer
-				systemUptime, uptimeErr := uptime.Get()
-				if uptimeErr != nil {
-					log.Warn().Err(uptimeErr).Msg("failed to get system uptime, using 0")
-					systemUptime = 0
-				}
-				monotonicStart := int64(systemUptime.Seconds())
-
-				// Determine clock reliability and source
-				clockReliable := helpers.IsClockReliable(now)
-				var clockSource string
-				if clockReliable {
-					clockSource = helpers.ClockSourceSystem
-				} else {
-					clockSource = helpers.ClockSourceEpoch
-				}
-
-				entry := &database.MediaHistoryEntry{
-					ID:             uuid.New().String(),
-					StartTime:      activeMedia.Started,
-					SystemID:       activeMedia.SystemID,
-					SystemName:     activeMedia.SystemName,
-					MediaPath:      activeMedia.Path,
-					MediaName:      activeMedia.Name,
-					LauncherID:     activeMedia.LauncherID,
-					PlayTime:       0,
-					BootUUID:       t.st.BootUUID(),
-					MonotonicStart: monotonicStart,
-					DurationSec:    0,
-					WallDuration:   0,
-					TimeSkewFlag:   false,
-					ClockReliable:  clockReliable,
-					ClockSource:    clockSource,
-					CreatedAt:      now,
-					UpdatedAt:      now,
-				}
-				dbid, addErr := t.db.UserDB.AddMediaHistory(entry)
-				if addErr != nil {
-					log.Error().Err(addErr).Msg("failed to add media history entry")
-				} else {
-					t.mu.Lock()
-					t.currentHistoryDBID = dbid
-					t.currentMediaStartTime = activeMedia.Started
-					t.currentMediaStartTimeMono = nowMono
-					t.mu.Unlock()
-					log.Debug().Int64("dbid", dbid).Msg("created media history entry")
-				}
-			}
-
-		case models.NotificationStopped:
-			// Media stopped - close history entry
-			t.mu.Lock()
-			dbid := t.currentHistoryDBID
-			startTime := t.currentMediaStartTime
-			startTimeMono := t.currentMediaStartTimeMono
-			t.currentHistoryDBID = 0
-			t.currentMediaStartTime = time.Time{}
-			t.currentMediaStartTimeMono = time.Time{}
-			t.mu.Unlock()
-
-			if dbid != 0 {
-				endTime := t.clock.Now()
-
-				// Calculate duration - prefer monotonic if available, fall back to wall-clock
-				var playTime int
-				if !startTimeMono.IsZero() {
-					// Use monotonic clock (more accurate, handles sleep)
-					endTimeMono := time.Now()
-					playTime = int(endTimeMono.Sub(startTimeMono).Seconds())
-				} else {
-					// Fall back to wall-clock (for tests or if mono not initialized)
-					playTime = int(endTime.Sub(startTime).Seconds())
-				}
-
-				closeErr := t.db.UserDB.CloseMediaHistory(dbid, endTime, playTime)
-				if closeErr != nil {
-					log.Error().Err(closeErr).Int64("dbid", dbid).Msg("failed to close media history entry")
-				} else {
-					log.Debug().Int64("dbid", dbid).Int("playTime", playTime).Msg("closed media history entry")
-				}
-			}
-		}
-	}
-}
-
-// updatePlayTime periodically updates the PlayTime for the currently active media
-// history entry every minute.
-func (t *mediaHistoryTracker) updatePlayTime(ctx context.Context) {
-	ticker := t.clock.NewTicker(1 * time.Minute)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-ticker.Chan():
-			t.mu.RLock()
-			dbid := t.currentHistoryDBID
-			startTime := t.currentMediaStartTime
-			startTimeMono := t.currentMediaStartTimeMono
-			t.mu.RUnlock()
-
-			if dbid != 0 {
-				// Calculate duration - prefer monotonic if available, fall back to wall-clock
-				var playTime int
-				switch {
-				case !startTimeMono.IsZero():
-					// Use monotonic clock (more accurate, handles sleep/hibernate)
-					nowMono := time.Now()
-					playTime = int(nowMono.Sub(startTimeMono).Seconds())
-				case !startTime.IsZero():
-					// Fall back to wall-clock (for tests or if mono not initialized)
-					playTime = int(t.clock.Since(startTime).Seconds())
-				default:
-					// No valid start time - skip update
-					continue
-				}
-
-				updateErr := t.db.UserDB.UpdateMediaHistoryTime(dbid, playTime)
-				if updateErr != nil {
-					log.Warn().Err(updateErr).Msg("failed to update media history play time")
-				} else {
-					log.Debug().Int64("dbid", dbid).Int("playTime", playTime).Msg("updated media history play time")
-				}
-			}
-		case <-ctx.Done():
-			return
-		}
-	}
-}
-
-// startPublishers initializes and starts all configured publishers.
-// Returns a slice of active publishers and a cancel function for graceful shutdown.
-func startPublishers(
-	st *state.State,
-	cfg *config.Instance,
-	notifChan <-chan models.Notification,
-) ([]publishers.Publisher, context.CancelFunc) {
-	activePublishers := make([]publishers.Publisher, 0)
-
-	mqttConfigs := cfg.GetMQTTPublishers()
-	if len(mqttConfigs) > 0 {
-		for _, mqttCfg := range mqttConfigs {
-			// Skip if explicitly disabled (nil = enabled by default)
-			if mqttCfg.Enabled != nil && !*mqttCfg.Enabled {
-				continue
-			}
-
-			log.Info().Msgf("starting MQTT publisher: %s (topic: %s)", mqttCfg.Broker, mqttCfg.Topic)
-
-			publisher := publishers.NewMQTTPublisher(mqttCfg.Broker, mqttCfg.Topic, mqttCfg.Filter)
-			if err := publisher.Start(st.GetContext()); err != nil {
-				log.Error().Err(err).Msgf("failed to start MQTT publisher for %s", mqttCfg.Broker)
-				continue
-			}
-
-			activePublishers = append(activePublishers, publisher)
-		}
-	}
-
-	for _, pcCfg := range cfg.GetPixelCadePublishers() {
-		if pcCfg.Enabled != nil && !*pcCfg.Enabled {
-			continue
-		}
-
-		log.Info().Msgf("starting PixelCade publisher: %s:%d", pcCfg.Host, pcCfg.Port)
-
-		publisher := publishers.NewPixelCadePublisher(
-			pcCfg.Host, pcCfg.Port, pcCfg.Mode, pcCfg.Filter,
-		)
-		if err := publisher.Start(st.GetContext()); err != nil {
-			log.Error().Err(err).Msgf("failed to start PixelCade publisher for %s", pcCfg.Host)
-			continue
-		}
-
-		activePublishers = append(activePublishers, publisher)
-	}
-
-	if len(activePublishers) > 0 {
-		log.Info().Msgf("started %d publisher(s)", len(activePublishers))
-	}
-
-	// CRITICAL: Always start the drain goroutine, even if there are no active publishers.
-	// The notifChan MUST be consumed or it will fill up and block the notification system.
-	// If there are no publishers, notifications are simply discarded after being consumed.
-	ctx, cancel := context.WithCancel(st.GetContext()) //nolint:gosec // G118: cancel returned to caller
-	go func() {
-		for {
-			select {
-			case <-ctx.Done():
-				log.Debug().Msg("publisher fan-out: stopping")
-				return
-			case notif, ok := <-notifChan:
-				if !ok {
-					log.Debug().Msg("publisher fan-out: notification channel closed")
-					return
-				}
-				// Publish to all active publishers sequentially
-				// If no publishers, notification is simply discarded
-				// Timeout in Publish() prevents blocking indefinitely
-				for _, pub := range activePublishers {
-					if err := pub.Publish(notif); err != nil {
-						log.Warn().Err(err).Msgf("failed to publish %s notification", notif.Method)
-					}
-				}
-			}
-		}
-	}()
-
-	return activePublishers, cancel
-}
-
-// checkAndResumeIndexing checks if media indexing was interrupted and automatically resumes it
-func checkAndResumeIndexing(
-	pl platforms.Platform,
-	cfg *config.Instance,
-	db *database.Database,
-	st *state.State,
-	pauser *syncutil.Pauser,
-) {
-	// Check if indexing was interrupted
-	indexingStatus, err := db.MediaDB.GetIndexingStatus()
-	if err != nil {
-		log.Debug().Err(err).Msg("failed to get indexing status during startup check")
-		return
-	}
-
-	// Only resume if indexing was interrupted (running or pending states)
-	if indexingStatus != mediadb.IndexingStatusRunning && indexingStatus != mediadb.IndexingStatusPending {
-		log.Debug().Msgf("indexing status is '%s', no auto-resume needed", indexingStatus)
-		return
-	}
-
-	log.Info().Msg("detected interrupted media indexing, automatically resuming")
-
-	// Get the systems that were being indexed from the database
-	// If not available, fall back to all systems
-	var systems []systemdefs.System
-	storedSystemIDs, err := db.MediaDB.GetIndexingSystems()
-	if err != nil || len(storedSystemIDs) == 0 {
-		log.Debug().Msgf("no stored systems found (err=%v, len=%d), defaulting to all systems",
-			err, len(storedSystemIDs))
-		systems = systemdefs.AllSystems()
-	} else {
-		// Convert system IDs to System objects
-		systems = make([]systemdefs.System, 0, len(storedSystemIDs))
-		for _, systemID := range storedSystemIDs {
-			if system, exists := systemdefs.Systems[systemID]; exists {
-				systems = append(systems, system)
-			} else {
-				log.Warn().Msgf("stored system ID '%s' not found in system definitions, skipping", systemID)
-			}
-		}
-		// If we couldn't resolve any systems, fall back to all systems
-		if len(systems) == 0 {
-			log.Warn().Msg("could not resolve any stored systems, falling back to all systems")
-			systems = systemdefs.AllSystems()
-		}
-	}
-
-	// Resume using the proper function with full notification support
-	// GenerateMediaDB spawns its own goroutine and returns immediately
-	err = methods.GenerateMediaDB(st.GetContext(), pl, cfg, st.Notifications, systems, db, pauser)
-	if err != nil {
-		log.Error().Err(err).Msg("failed to start auto-resume of media indexing")
-	}
-}
-
-// checkAndResumeOptimization checks if optimization was interrupted and automatically resumes it
-func checkAndResumeOptimization(db *database.Database, ns chan<- models.Notification, pauser *syncutil.Pauser) {
-	status, err := db.MediaDB.GetOptimizationStatus()
-	if err != nil {
-		log.Debug().Err(err).Msg("failed to get optimization status during startup check")
-		return
-	}
-
-	// Resume if optimization was interrupted or failed
-	if status == mediadb.IndexingStatusPending ||
-		status == mediadb.IndexingStatusRunning ||
-		status == mediadb.IndexingStatusFailed {
-		log.Info().Msgf("detected incomplete optimization (status: %s), automatically resuming", status)
-		db.MediaDB.RunBackgroundOptimization(func(optimizing bool) {
-			notifications.MediaIndexing(ns, models.IndexingStatusResponse{
-				Exists:     true,
-				Indexing:   false,
-				Optimizing: optimizing,
-			})
-		}, pauser)
-	} else {
-		log.Debug().Msgf("optimization status is '%s', no auto-resume needed", status)
-	}
 }
